requirements.txt
###### Requirements without Version Specifiers ######`
flask
flask_sqlalchemy
pandas
smartsheet-python-sdk
yfinance
googlemaps